refactor: remove re-export shims from auth modules
Update all imports to use new modular paths directly:
- @auth/serverSideKeyAccess → @auth/serverKeys
- @auth/tierModelAccess → @auth/tier

Delete obsolete re-export files that were only providing
backward compatibility during the transition.
